apps/ai-brain/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
numpy==1.26.4
pandas==2.2.1
openai==1.13.3
scikit-learn==1.4.1
redis==5.0.1
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
